Add files via upload
This file named "CONTRIBUTING.md" is an essential governance document that needs to be incorporated in the root directory. It describes how to contribute to the RuFaS program through 5 contribution categories. It also provides links to the code of conduct and other information related to coding conventions, style guidance, versioning policy, review process for code contributions, and guidance for non-coding contributions.